+++ frontend/components/AnalyticsManager.php
@@ -27,11 +27,15 @@ class AnalyticsManager {
         // ['user', 'age'],
         // ['user', 'gender'],
         // Message analytics
-        // ['message', 'm:from_users'],
-        // ['message', 'm:responses'],
-        // ['message', 'm:notifications'],
+        ['message', 'requests'],
+        ['message', 'responses'],
+        ['message', 'notifications'],
         // Task analytics
+        ['task', 'new'],
+        ['task', 'active'],
+        ['task', 'closed'],
         // Transaction analytics
+        ['transaction', 'total'],
         // Badge analytics
     ];
 
@@ -42,8 +46,10 @@ class AnalyticsManager {
         ['user', 'age'],
         ['user', 'gender'],
         // Message analytics
+        ['message', 'all'],
         // Task analytics
         // Transaction analytics
+        ['transaction', 'label'],
         // Badge analytics
     ];
 
@@ -159,47 +165,60 @@ class AnalyticsManager {
         ];
     }
 
-    # TODO
     private function messageData($appId, $timespan) {
+        $segmentationData = $this->get($appId, 'message', 'all', $timespan)->content();
         return [
-            'platform' => [
+            'requests' => [
                 'total' => null, # TODO
-                'period' => $this->get($appId, 'message', 'notifications', $timespan)->result->count,
+                'period' => $this->get($appId, 'message', 'requests', $timespan)->content(),
             ],
-            'app' => [
+            'responses' => [
                 'total' => null, # TODO
-                'period' => $this->get($appId, 'message', 'responses', $timespan)->result->count,
+                'period' => $this->get($appId, 'message', 'responses', $timespan)->content(),
             ],
-            'users' => [
+            'notifications' => [
                 'total' => null, # TODO
-                'period' => $this->get($appId, 'message', 'from_users', $timespan)->result->count,
-            ]
+                'period' => $this->get($appId, 'message', 'notifications', $timespan)->content(),
+            ],
+            'segmentation' => $segmentationData ? $segmentationData : [],
         ];
     }
 
-    # TODO
     private function taskData($appId, $timespan) {
         return [
-            'total' => 101,
-            'new' => 21,
-            'active' => 86,
-            'closed' => 15,
+            'new' => [
+                'total' => null, # TODO
+                'period' => $this->get($appId, 'task', 'new', $timespan)->content(),
+            ],
+            'active' => [
+                'total' => null, # TODO
+                'period' => $this->get($appId, 'task', 'active', $timespan)->content(),
+            ],
+            'closed' => [
+                'total' => null, # TODO
+                'period' => $this->get($appId, 'task', 'closed', $timespan)->content(),
+            ],
         ];
     }
 
-    # TODO
     private function transactionData($appId, $timespan) {
+        $segmentationData = $this->get($appId, 'transaction', 'label', $timespan)->content();
+        $newTransactions = $this->get($appId, 'transaction', 'total', $timespan)->content();
+
+        # This is a fix required to remove the `CREATE_TASK` label. This is not a real transaction
+        # label; it is, in fact, a control variable used inside the Task Manager.
+        if (is_array($segmentationData) && array_key_exists('CREATE_TASK', $segmentationData)) {
+            $createTaskTransactions = $segmentationData['CREATE_TASK'];
+            $newTransactions -= $createTaskTransactions;
+            unset($segmentationData['CREATE_TASK']);
+        }
+
         return [
-            'total' => 111,
-            'new' => 81,
-            'distribution' => [
-                'answerTransaction' => 5,
-                'notAnswerTransaction' => 10,
-                'bestAnswerTransaction' => 15,
-                'moreAnswerTransaction' => 20,
-                'reportQuestionTransaction' => 25,
-                'reportAnswerTransaction' => 6
-            ]
+            'new' => [
+                'total' => null, # TODO
+                'period' => $newTransactions,
+            ],
+            'segmentation' => $segmentationData ? $segmentationData : [],
         ];
     }
 
@@ -214,9 +233,9 @@ class AnalyticsManager {
     public function prepareData($appId, $timespan) {
         $data = [];
         $data['users'] = $this->userData($appId, $timespan);
-        // $data['messages'] = $this->messageData($appId, $timespan);
-        // $data['tasks'] = $this->taskData($appId, $timespan);
-        // $data['transactions'] = $this->transactionData($appId, $timespan);
+        $data['messages'] = $this->messageData($appId, $timespan);
+        $data['tasks'] = $this->taskData($appId, $timespan);
+        $data['transactions'] = $this->transactionData($appId, $timespan);
         return $data;
     }