+++ frontend/views/developer/_stats.php
@@ -17,9 +17,9 @@
         </div>
 
         <?php echo Yii::$app->controller->renderPartial('stats/_users', ['statsData' => $statsData['users']]); ?>
-        <?php // echo Yii::$app->controller->renderPartial('stats/_messages', ['statsData' => $statsData['messages']]); ?>
-        <?php // echo Yii::$app->controller->renderPartial('stats/_tasks', ['statsData' => $statsData['tasks']]); ?>
-        <?php // echo Yii::$app->controller->renderPartial('stats/_transactions', ['statsData' => $statsData['transactions']]); ?>
+        <?php echo Yii::$app->controller->renderPartial('stats/_messages', ['statsData' => $statsData['messages']]); ?>
+        <?php echo Yii::$app->controller->renderPartial('stats/_tasks', ['statsData' => $statsData['tasks']]); ?>
+        <?php echo Yii::$app->controller->renderPartial('stats/_transactions', ['statsData' => $statsData['transactions']]); ?>
 
     </div>
 </div>